+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Features:
 * Don't show 403 if user can't change collaborators: show explanation, hide button
 * Add export of highlights to CSV and Excel (.xlsx) formats
 * Support PBKDF2 passwords (Taguette no longer requires 'bcrypt')
+* Set umask to 077 by default (and add corresponding command-line options)
 
 0.8 (2019-06-15)
 ----------------